% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dip21_details.R
\name{dip21_details}
\alias{dip21_details}
\title{function giving back details on a dip21_search request}
\usage{
dip21_details(xml)
}
\arguments{
\item{xml}{the dip21_search(res="content") result}
}
\description{
function giving back details on a dip21_search request
}

