% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dip21_search.R
\name{dip21_search}
\alias{dip21_search}
\title{doing the search}
\usage{
dip21_search(wahlperiode = 18, vorgangstyp = "Alle", nummer = NULL,
  suchwort = NULL, force = FALSE, res = c("nhits", "trefferliste"))
}
\arguments{
\item{wahlperiode}{legislative term use e.g. wahlperioden(18) to get the
right value to put into the form}

\item{vorgangstyp}{type of proceeding; use e.g. vorgangstypen("Gesetzgebung")
to get the right values to put in.}

\item{nummer}{"Drucksache" document number related to the proceeding}

\item{suchwort}{search term}

\item{force}{option passed along to init_search_page(), whether or not force
re-initialisation of seeerver session}

\item{res}{A character vector naming the results to be returned:
nhits - number of procedings found
npages - number of pages that show procedings
trefferliste - lsit of hits with some very basic information
content - the content of the response to the HTTP request send}
}
\description{
doing the search
}

