% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dip21_search.R
\name{n_search_results}
\alias{n_search_results}
\title{function extractiong number of search result hits}
\usage{
n_search_results(xml)
}
\arguments{
\item{xml}{the seach result to extract number from}
}
\description{
function extractiong number of search result hits
}

