% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dip21_search.R
\name{npages_search_results}
\alias{npages_search_results}
\title{function extracting number of search result hit pages}
\usage{
npages_search_results(xml)
}
\arguments{
\item{xml}{the seach result to extract number from}
}
\description{
function extracting number of search result hit pages
}

