% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dip21_search.R
\name{page_search_results}
\alias{page_search_results}
\title{function extracting current number of search result hit page}
\usage{
page_search_results(xml)
}
\arguments{
\item{xml}{the seach result to extract number from}
}
\description{
function extracting current number of search result hit page
}

