% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_scraping.R
\name{parse_and_fill_form}
\alias{parse_and_fill_form}
\title{parsing, filling out and sending form}
\usage{
parse_and_fill_form(xml, base_url = "", submit = NULL, ...)
}
\arguments{
\item{xml}{HTML/XML which contains the form. Forms will be parsed by rvest::html_form().}

\item{base_url}{the base url to send form to}

\item{submit}{which submit to use. See \link{submit_request}}
}
\description{
parsing, filling out and sending form
}

