% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_request.R
\name{submit_request}
\alias{submit_request}
\title{alternative submit_request to rvest's original submit request}
\usage{
submit_request(form, submit = NULL)
}
\description{
alternative submit_request to rvest's original submit request
}

