% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dip21_search.R
\name{trefferliste}
\alias{trefferliste}
\title{function for getting trefferliste from}
\usage{
trefferliste(xml)
}
\arguments{
\item{xml}{xml from which to extract form for send}
}
\description{
function for getting trefferliste from
}

