% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dip21_search.R
\name{vorgangstypen}
\alias{vorgangstypen}
\title{function returning options for tpe of proceeding}
\usage{
vorgangstypen(regex = NULL)
}
\arguments{
\item{vt}{a regular expression used to look up labels and return their values}
}
\description{
function returning options for tpe of proceeding
}

