% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dip21_search.R
\name{wahlperioden}
\alias{wahlperioden}
\title{function returning options for legislative terms}
\usage{
wahlperioden(regex = NULL)
}
\arguments{
\item{wp}{a regular expression used to look up labels and return their values}
}
\description{
function returning options for legislative terms
}

